Extracting linkparent tests to another file
